Make model delimiter more lenient to whitespace
Currently, there's an expectation to format the model in a specific way,
where every delimiter is `", "`. This rule isn't documented anywhere and
makes the file formatting inflexible and error-prone.

I suggest we make the delimiter a bit more flexible to whitespace around
the comma, so that it allows people to format their model CSV without
breaking the enforcer initialization.